Remove MC Object Class, ProductType Params, and MagicCore
